requirements.txt
mysql
python-telegram-bot